Add method to release job to the store adapter
This will allow us to remove a lock from the database by a console
command. The delete has to be manually because releasing the lock is
pretected against other processes but the owner.

However the command is meant to release te lock in case of emergency, so
no the lock protection is bypassed by intend